refactor:使用PostgreSQL代替了MySQL,同时给数据库增加了单独的一层

+++ pkg/category/dao/dao_interface.go
@@ -7,10 +7,10 @@ import (
 
 type CategoryDaoInfertace interface {
 	UpdateCategoryType(ctx context.Context, category *model.Category) error
-	GetCategoryById(ctx context.Context, id int64) (*model.Category, error)
+	GetCategoryById(ctx context.Context, id uint) (*model.Category, error)
 	GetCategoryPage(ctx context.Context, name string, page, pageSize, categoryType int) (categories []model.Category, total int64, err error)
-	ChangeCategoryStatus(ctx context.Context, id int64, status int) error
+	ChangeCategoryStatus(ctx context.Context, id uint, status int) error
 	CreateCategory(ctx context.Context, category *model.Category) error
-	DeleteCategory(ctx context.Context, id int64) error
+	DeleteCategory(ctx context.Context, id uint) error
 	GetCategoryByType(ctx context.Context, categoryType int) ([]model.Category, error)
 }
