+++ pkg/dish/DTO/dish.go
@@ -10,7 +10,7 @@ type UpdateDishResponse struct {
 }
 
 type DeleteDishRequest struct {
-	IDs []int64 `form:"ids" binding:"required"` // ids
+	IDs []uint `form:"ids" binding:"required"` // ids
 }
 
 type DeleteDishResponse struct {
@@ -24,7 +24,7 @@ type AddDishResponse struct {
 }
 
 type SearchDishByIDRequest struct {
-	ID int64 `uri:"id" binding:"required"` // 菜品id
+	ID uint `uri:"id" binding:"required"` // 菜品id
 }
 
 type SearchDishByIDResponse struct {
@@ -32,7 +32,7 @@ type SearchDishByIDResponse struct {
 }
 
 type SearchDishByCategoryRequest struct {
-	CategoryID int64 `uri:"category_id" binding:"required"` // 分类id
+	CategoryID uint `uri:"category_id" binding:"required"` // 分类id
 }
 
 type SearchDishByCategoryResponse struct {
@@ -40,7 +40,7 @@ type SearchDishByCategoryResponse struct {
 }
 
 type SearchDishByPageRequest struct {
-	CategoryID int64  `form:"category_id"`                 // 分类id
+	CategoryID uint   `form:"category_id"`                 // 分类id
 	Name       string `form:"name"`                        // 菜品名称
 	Status     int    `form:"status"`                      // 菜品状态
 	Page       int    `form:"page" binding:"required"`     // 页码
@@ -53,8 +53,8 @@ type SearchDishByPageResponse struct {
 }
 
 type ChangeDishStatusRequest struct {
-	Status int   `uri:"status"` // 菜品状态，1表示上架，0表示下架
-	ID     int64 `form:"id"`    // 菜品id
+	Status int  `uri:"status"` // 菜品状态，1表示上架，0表示下架
+	ID     uint `form:"id"`    // 菜品id
 }
 
 type ChangeDishStatusResponse struct {
